+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - XXXX-XX-XX
+
+### Changed
+- Changed to use Tailwind 1.2’s new plugin definition syntax
+
 ## [1.0.1] - 2019-12-20
 
 ### Fixed
@@ -15,6 +20,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-alt/compare/v1.0.1...HEAD
-[1.0.1]: https://github.com/benface/tailwindcss-alt/releases/tag/v1.0.1
+[Unreleased]: https://github.com/benface/tailwindcss-alt/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/benface/tailwindcss-alt/compare/v1.0.1...v2.0.0
+[1.0.1]: https://github.com/benface/tailwindcss-alt/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/benface/tailwindcss-alt/releases/tag/v1.0.0